% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_fat.R
\docType{data}
\name{body_fat}
\alias{body_fat}
\title{Body Fat Prediction Dataset}
\format{
A data frame with 252 rows and 15 covariate variables and 1 response variable
}
\source{
\url{https://www.kaggle.com/datasets/fedesoriano/body-fat-prediction-dataset}
}
\description{
Lists estimates of the percentage of body fat determined by underwater
weighing and various body circumference measurements for 252 men. 
Accurate measurement of body fat is inconvenient/costly and it is desirable to have easy methods of estimating body fat that are not inconvenient/costly.
}
\details{
The variables listed below, from left to right, are:
\itemize{
\item Density determined from underwater weighing
\item Age (years)
\item Weight (lbs)
\item Height (inches)
\item Neck circumference (cm)
\item Chest circumference (cm)
\item Abdomen 2 circumference (cm)
\item Hip circumference (cm)
\item Thigh circumference (cm)
\item Knee circumference (cm)
\item Ankle circumference (cm)
\item Biceps (extended) circumference (cm)
\item Forearm circumference (cm)
\item Wrist circumference (cm)
}
}
\examples{
data(body_fat)
set.seed(221212)
train = sample(1:252,100)
train_data = data.frame(body_fat[train,])
test_data = data.frame(body_fat[-train,])

library(ODRF)

rf = ODRF(Density~.,train_data,type='regression')
pred <- predict(rf,test_data[,-1],weight = FALSE)$prediction
#estimation error
mean((pred-test_data[,1])^2)

tree = ODT(Density~.,train_data,type='regression')
pred <- predict(tree,test_data[,-1])
#estimation error
mean((pred-test_data[,1])^2)
}
\references{
Bailey, Covert (1994). Smart Exercise: Burning Fat, Getting Fit, Houghton-Mifflin Co., Boston, pp. 179-186.

Behnke, A.R. and Wilmore, J.H. (1974). Evaluation and Regulation of Body Build and Composition, Prentice-Hall, Englewood Cliffs, N.J.

Siri, W.E. (1956), "Gross composition of the body", in Advances in Biological and Medical Physics, vol. IV, edited by J.H. Lawrence and C.A. Tobias, Academic Press, Inc., New York.

Katch, Frank and McArdle, William (1977). Nutrition, Weight Control, and Exercise, Houghton Mifflin Co., Boston.

Wilmore, Jack (1976). Athletic Training and Physical Fitness: Physiological Principles of the Conditioning Process, Allyn and Bacon, Inc., Boston.
}
\keyword{datasets}
