% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{defaults}
\alias{defaults}
\title{Default values passed to RotMat*}
\usage{
defaults(
  paramList,
  type = "i-classification",
  dimX = NULL,
  weights = NULL,
  catLabel = NULL
)
}
\arguments{
\item{paramList}{a list (possibly empty), to be populated with a set
of default values to be passed to a \code{RotMat*} function.}

\item{type}{The criterion used for splitting the variable. 'i-classification': information gain (classification, default), 
'g-classification': gini impurity index (classification) or 'regression': mean square error (regression).}

\item{weights}{A vector of length same as \code{data} that are positive weights.(default NULL)}

\item{catLabel}{A category labels of class \code{list} in prediction variables. (for details see Details and Examples)}

\item{NodeRotateFun}{Name of the function of class \code{character} that implements a linear combination of predictor variables in the split node. Default is "RotMatPPO" with model="PPR". (see \code{\link{RotMatPPO}}) 
Users can define this function, for details see \code{\link{RotMatMake}}.}

\item{ncolX}{an integer denoting the number of columns in the design matrix X.}
}
\value{
Default parameters of the RotMat* function
\itemize{
\item \code{dimX} an integer denoting the number of columns in the design matrix X.
\item \code{dimProj} Number of variables to be projected, default dimProj="Rand": random from 1 to ncol(X).
\item \code{numProj} the number of desired columns in the projection matrix, is also the number of projection directions.(default ceiling(sqrt(dimX)))
\item \code{catLabel} A category labels of class \code{list} in prediction variables，for details see Examples of \code{\link{ODRF}}.
\item \code{weights} A vector of length same as \code{data} that are positive weights.(default NULL)
\item \code{lambda} Parameter of the Poisson distribution (default 1).
\item \code{sparsity} #' @param sparsity a real number in \eqn{(0,1)} that specifies the distribution of non-zero elements in the random matrix. 
When sparsity="pois" means that non-zero elements are generated by the p(\code{lambda}) Poisson distribution.
\item \code{prob} a probability \eqn{\in (0,1)} used for sampling from.
\item \code{randDist} Parameter of the Poisson distribution (default 1).
\item \code{type} type The criterion used for splitting the variable. 'i-classification': information gain (classification, default), 
'g-classification': gini impurity index (classification) or 'regression': mean square error (regression).
\item \code{model} model for projection pursuit. (see \code{\link{PPO}})
}
}
\description{
Given the parameter list and the categorical map this function
populates the values of the parameter list accoding to our 'best'
known general use case parameters.
}
\examples{
set.seed(1)
paramList <- list(dimX = 8, numProj= 3, sparsity=0.25, prob=0.5)
(paramList=defaults(paramList,type='i-classification'))
}
\seealso{
\code{\link{RotMatMake}} \code{\link{RotMatRand}} \code{\link{RotMatRF}} \code{\link{RotMatMake}}
}
\keyword{internal}
