% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODRF_error.R
\name{ODRF.error}
\alias{ODRF.error}
\title{oblique decision random forest error}
\usage{
ODRF.error(ppForest, data, newdata = NULL, ...)
}
\arguments{
\item{ppForest}{an object of class \code{ODRF}, as that created by the function \code{\link{ODRF}}.}

\item{data}{Training data of class \code{data.frame} in which to interpret the variables named in the formula.If data is missing it is obtained from the current environment by \code{formula}.}

\item{newdata}{A data frame or matrix containing new data.}
}
\value{
OOB error and test error, classification error rate for classification or RPE(MSE/mean((ytest-mean(y))^2)) for regression.
}
\description{
the error of class \code{ODRF} at different number of trees.
}
\examples{
library(ODRF)

data(seeds)
set.seed(221212)
train = sample(1:209,100)
train_data = data.frame(seeds[train,])
test_data = data.frame(seeds[-train,])

forest = ODRF(varieties_of_wheat~.,train_data,type='i-classification')
error=ODRF.error(forest,train_data,test_data)

}
\seealso{
\code{ODRF} \code{plot.ODRF.error}
}
