% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_ODRF.R, R/online_ODT.R
\name{online.ODRF}
\alias{online.ODRF}
\alias{online.ODT}
\title{A machine learning algorithm for training class \code{ODRF}.}
\usage{
\method{online}{ODRF}(ppForest, data, weights = NULL)

\method{online}{ODRF}(ppTree, data, weights = NULL)
}
\arguments{
\item{data}{Training data of class \code{data.frame} is used to update the object of class \code{ODT}.}

\item{weights}{A vector of length same as \code{data} that are positive weights. (default NULL)}

\item{obj}{an object of class \code{ODT}.}
}
\value{
The same result as \code{ODRF}.

The same result as \code{ODT}.
}
\description{
The \code{\link{ODRF}}is constantly updated by multiple batches of data to optimize the model.

The \code{\link{ODT}}is constantly updated by multiple batches of data to optimize the model.
}
\examples{
#Classification with Oblique Decision Tree
data(seeds)
set.seed(221212)
train = sample(1:209,100)
train_data = data.frame(seeds[train,])
test_data = data.frame(seeds[-train,])

tree = ODRF(varieties_of_wheat~.,train_data[seq(floor(nrow(train_data)/2)),],type='i-classification')
tree = online(tree,train_data[-seq(floor(nrow(train_data)/2)),])
pred <- predict(tree,test_data[,-8])
#estimation error
(mean(pred!=test_data[,8]))

#Regression with Oblique Decision Tree
data(body_fat)
set.seed(221212)
train = sample(1:252,100)
train_data = data.frame(body_fat[train,])
test_data = data.frame(body_fat[-train,])

tree = ODRF(Density~.,train_data[seq(floor(nrow(train_data)/2)),],type='regression')
tree = online(tree,train_data[-seq(floor(nrow(train_data)/2)),])
pred <- predict(tree,test_data[,-8])
#estimation error
mean((pred-test_data[,1])^2)

#Classification with Oblique Decision Tree
data(seeds)
set.seed(221212)
train = sample(1:209,100)
train_data = data.frame(seeds[train,])
test_data = data.frame(seeds[-train,])

tree = ODT(varieties_of_wheat~.,train_data[seq(floor(nrow(train_data)/2)),],type='i-classification')
tree = online(tree,train_data[-seq(floor(nrow(train_data)/2)),])
pred <- predict(tree,test_data[,-8])
#estimation error
(mean(pred!=test_data[,8]))

#Regression with Oblique Decision Tree
data(body_fat)
set.seed(221212)
train = sample(1:252,100)
train_data = data.frame(body_fat[train,])
test_data = data.frame(body_fat[-train,])

tree = ODT(Density~.,train_data[seq(floor(nrow(train_data)/2)),],type='regression')
tree = online(tree,train_data[-seq(floor(nrow(train_data)/2)),])
pred <- predict(tree,test_data[,-8])
#estimation error
mean((pred-test_data[,1])^2)

}
\seealso{
\code{\link{ODRF}} \code{\link{prune.ODRF}}

\code{\link{ODT}} \code{\link{prune.ODT}}
}
