% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online.R
\name{online}
\alias{online}
\title{online structure learning for class \code{ODT} and \code{ODRF}.}
\usage{
online(obj, ...)
}
\arguments{
\item{obj}{an object of class \code{ODT} or \code{ODRF}.}

\item{...}{For other parameters related to class \code{obj}, see \code{ODT} or \code{ODRF}.}
}
\value{
object of class \code{ODT} or \code{ODRF}.
}
\description{
\code{\link{ODT}} and \code{\link{ODRF}} are constantly updated by multiple batches of data to optimize the model. \code{online} is a S3 method for class \code{ODT} and \code{ODRF}.
}
\seealso{
\code{\link{ODT}} \code{\link{ODRF}} \code{\link{online.ODT}} \code{\link{online.ODRF}}
}
