% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ODRF_error.R
\name{plot.ODRF.error}
\alias{plot.ODRF.error}
\title{Plot method for ODRF objects}
\usage{
\method{plot}{ODRF.error}(
  Err,
  lty = 1,
  main = paste0("Oblique ", ifelse(Err$type == "regression", "Regression",
    "Classification"), " Forest"),
  ...
)
}
\arguments{
\item{Err}{Object of class \code{\link{ODRF.error}}.}

\item{lty}{a vector of line types, see \code{\link{par}}.}

\item{main}{main title of the plot.}

\item{...}{arguments to be passed to methods.}
}
\value{
OOB error and test error, classification error rate for classification or RPE(MSE/mean((ytest-mean(y))^2)) for regression.
}
\description{
Draw the error graph of class \code{ODRF} at different number of trees.
}
\examples{
library(ODRF)

data(seeds)
set.seed(221212)
train = sample(1:209,100)
train_data = data.frame(seeds[train,])
test_data = data.frame(seeds[-train,])

forest = ODRF(varieties_of_wheat~.,train_data,type='i-classification')
error=ODRF.error(forest,train_data,test_data)
plot(error)

}
\seealso{
\code{ODRF} \code{ODRF.error}
}
\keyword{forest}
