% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prune_ODT.R
\name{plot.prune.ODT}
\alias{plot.prune.ODT}
\title{PPtree plot}
\usage{
\method{plot}{prune.ODT}(
  pptree,
  position = "bottomleft",
  main = paste0("Oblique ", ifelse(pptree$method == "regression", "Regression",
    "Classification"), " Tree"),
  ...
)
}
\arguments{
\item{main}{main title}

\item{...}{arguments to be passed to methods}

\item{x}{PPtreeclass object}

\item{font.size}{font size of plot}

\item{width.size}{size of eclipse in each node.}

\item{sub}{sub title}
}
\description{
projection pursuit classification tree plot
}
\details{
Draw projection pursuit classification tree with tree structure. It is 
modified from a function in party library.
}
\examples{
data(iris)
class(pptree)="prune.ODT"
Tree.result <- PPTreeclass(Species~., data = iris,"LDA")
Tree.result
plot(Tree.result,xjust=3)
}
\references{
Lee, EK(2017) 
PPtreeViz: An R Package for Visualizing Projection Pursuit Classification 
Trees, Journal of Statistical Software <doi:10.18637/jss.v083.i08>
}
\keyword{tree}
