% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ODT_depth.R
\name{plot_ODT_depth}
\alias{plot_ODT_depth}
\title{oblique decision tree depth plot}
\usage{
plot_ODT_depth(
  formula,
  data = NULL,
  newdata = NULL,
  type = "i-classification",
  NodeRotateFun = "RotMatPPO",
  paramList = NULL,
  main = paste0("Oblique ", ifelse(type == "regression", "Regression",
    "Classification"), " Tree"),
  ...
)
}
\arguments{
\item{formula}{Object of class \code{formula} with a response but no interaction terms describing the model to fit. If this is a data frame, it is taken as the model frame. (see \code{\link{model.frame}})}

\item{data}{Training data of class \code{data.frame} in which to interpret the variables named in the formula.If data is missing it is obtained from the current environment by \code{formula}.}

\item{newdata}{A data frame or matrix containing new data.}

\item{type}{The criterion used for splitting the nodes. g-classification': gini impurity index(default) and i-classification': information gain for classification; 'regression': mean square error for regression.}

\item{NodeRotateFun}{Name of the function of class \code{character} that implements a linear combination of predictor variables in the split node. Default is "RotMatPPO" with model="PPR". (see \code{\link{RotMatPPO}}) 
Users can define this function, for details see \code{\link{RotMatMake}}.}

\item{paramList}{Parameters in a named list to be used by \code{ODT}.}

\item{main}{main title}

\item{...}{arguments to be passed to methods.}
}
\value{
OOB error and test error, classification error rate for classification or RPE(MSE/mean((ytest-mean(y))^2)) for regression.
}
\description{
Draw the error graph of class \code{ODT} at different depths.
}
\examples{
data(body_fat)
set.seed(221212)
train = sample(1:252,100)
train_data = data.frame(body_fat[train,])
test_data = data.frame(body_fat[-train,])
plot_ODT_depth(Density~.,train_data,test_data,type='regression')

}
\seealso{
\code{\link{ODT}} \code{\link{plot.ODT}}
}
\keyword{tree}
