% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ODRF.R
\name{print.ODRF}
\alias{print.ODRF}
\title{print ODRF}
\usage{
\method{print}{ODRF}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{ODRF}}.}

\item{...}{Arguments to be passed to methods.}
}
\value{
OOB error, misclassification rate (MR) for classification or mean square error (MSE) for regression.
}
\description{
Print contents of ODRF object.
}
\examples{
data(iris)
forest <- ODRF(Species ~ ., data = iris, parallel = FALSE, ntrees = 50)
forest

}
\seealso{
\code{\link{ODRF}}
}
\keyword{forest}
\keyword{print}
