% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ODT.R
\name{print.ODT}
\alias{print.ODT}
\title{Print ODT result}
\usage{
\method{print}{ODT}(ppTree, projection = FALSE, cutvalue = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{ppTree}{an object of class \code{\link{ODT}}.}

\item{projection}{print projection coefficients in each node if TRUE.}

\item{cutvalue}{print cutoff values in each node if TRUE.}

\item{verbose}{print if TRUE, no output if FALSE.}

\item{...}{arguments to be passed to methods.}
}
\description{
Print the oblique decision tree result
}
\examples{
data(iris)
tree <- ODT(Species~.,data = iris)
tree
print(tree,projection=TRUE,cutvalue=TRUE)

}
\references{
Lee, EK(2017) 
PPtreeViz: An R Package for Visualizing Projection Pursuit Classification 
Trees, Journal of Statistical Software <doi:10.18637/jss.v083.i08>
}
\seealso{
\code{\link{ODT}}
}
\keyword{tree}
