% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_ODT.R
\name{prune.ODT}
\alias{prune.ODT}
\title{Pruning of class \code{ODT}.}
\usage{
\method{prune}{ODT}(ppTree, data, MaxDepth = 1)
}
\arguments{
\item{data}{validation data of class \code{data.frame} is used to prune the object of class \code{ODT}.}

\item{MaxDepth}{The maximum depth of the tree after pruning. (Default 1)}

\item{obj}{an object of class \code{\link{ODT}}.}
}
\value{
\itemize{an object of class \code{ODT} and \code{prune.ODT}. 
\item{\code{ppTree} The same result as \code{ODT}.}
\item{\code{pruneError} Error of validation data after each pruning, classification error rate for classification or RPE(MSE/mean((ytest-mean(y))^2)) for regression.}
}
}
\description{
Prune \code{ODT} from bottom to top with validation data based on prediction error.
}
\examples{
library(ODRF)

#Classification with Oblique Decision Tree
data(seeds)
set.seed(221212)
train = sample(1:209,100)
train_data = data.frame(seeds[train,])
test_data = data.frame(seeds[-train,])

tree = ODT(varieties_of_wheat~.,train_data[seq(floor(nrow(train_data)/2)),],type='i-classification')
tree = prune(tree,train_data[-seq(floor(nrow(train_data)/2)),])
pred <- predict(tree,test_data[,-8])
#estimation error
(mean(pred!=test_data[,8]))

#Regression with Oblique Decision Tree
data(body_fat)
set.seed(221212)
train = sample(1:252,100)
train_data = data.frame(body_fat[train,])
test_data = data.frame(body_fat[-train,])

tree = ODT(Density~.,train_data[seq(floor(nrow(train_data)/2)),],type='regression')
tree = prune(tree,train_data[-seq(floor(nrow(train_data)/2)),])
pred <- predict(tree,test_data[,-8])
#estimation error
mean((pred-test_data[,1])^2)

}
\seealso{
\code{ODT} \code{plot.prune.ODRF}
}
