% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Pruning of class \code{ODT} or \code{ODRF}.}
\usage{
prune(obj, ...)
}
\arguments{
\item{obj}{an object of class \code{ODT} or \code{ODRF}.}

\item{...}{For other parameters related to class \code{obj}, see \code{\link{ODT}} or \code{\link{ODRF}}.}
}
\value{
an object of class \code{ODT} and \code{prune.ODT}.
}
\description{
Prune \code{ODT} or \code{ODRF} from bottom to top with validation data based on prediction error. \code{prune} is a S3 method for class class \code{ODT} and \code{ODRF}.
}
\seealso{
\code{ODT} \code{\link{ODRF}} \code{prune.ODT} \code{\link{prune.ODRF}}
}
