% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotMat.R
\name{RandRot}
\alias{RandRot}
\title{Samples a p x p uniformly random rotation matrix}
\usage{
RandRot(p)
}
\arguments{
\item{p}{The columns of an n by p numeric matrix or data frame.}
}
\value{
A p x p uniformly random rotation matrix.
}
\description{
Samples a p x p uniformly random rotation matrix via QR decomposition
of a matrix with elements sampled iid from a standard normal distribution.
}
\examples{
set.seed(220828)
(ODRF:::RandRot(100))

}
\seealso{
\code{\link{RotMatPPO}} \code{\link{RotMatRand}} \code{\link{RotMatRF}} \code{\link{RotMatMake}}
}
\keyword{internal}
\keyword{tree}
