% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotMat.R
\name{RotMatRF}
\alias{RotMatRF}
\title{Create a Projection Matrix: Random Forest (RF)}
\usage{
RotMatRF(dimX, numProj, catLabel = NULL, ...)
}
\arguments{
\item{dimX}{the number of dimensions.}

\item{numProj}{the number of desired columns in the projection matrix, is also the number of projection directions.(default ceiling(sqrt(dimX)))}

\item{catLabel}{A category labels of class \code{list} in prediction variables，for details see Examples of \code{\link{ODRF}}.}

\item{...}{used to handle superfluous arguments passed in using paramList.}
}
\value{
\itemize{A random matrix to use in running \code{\link{ODT}}.
\item{Variable: variables to be projected.} 
\item{Number: Number of variables to be projected.}
\item{Coefficient: Coefficients of the projection matrix.}
}
}
\description{
Create a projection matrix with coefficient 1, so that ODRF (ODT) has the same partition variables as random forest (tree).
}
\examples{
paramList <- list(dimX = 8, numProj= 3, catLabel=NULL)
set.seed(2)
(RotMat <- do.call(RotMatRF, paramList))
}
\seealso{
\code{\link{RotMatPPO}} \code{\link{RotMatRand}} \code{\link{RotMatMake}}
}
