% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_party_ODT.R
\name{as.party.ODT}
\alias{as.party.ODT}
\title{ODT as \code{party}}
\usage{
\method{as.party}{ODT}(ppTree, data, ...)
}
\arguments{
\item{ppTree}{an object of class \code{\link{ODT}}.}

\item{data}{Training data of class \code{data.frame} is used to convert the object of class \code{ODRF}. 
and \code{data} must be the training data \code{data} in \code{\link{ODT}}.}

\item{...}{arguments to be passed to methods}
}
\description{
Functions coercing \code{ODT} object to objects of class \code{party}.
}
\examples{
data(iris)
tree <- ODT(Species~.,data = iris)
tree
party.tree<- as.party(tree,data = iris)
party.tree
plot(party.tree)

}
\references{
Lee, EK(2017) PPtreeViz: An R Package for Visualizing Projection Pursuit Classification Trees, Journal of Statistical Software <doi:10.18637/jss.v083.i08>
}
\seealso{
\code{\link{OOT}} \code{\link{party}}
}
\keyword{tree}
