% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODRF_error.R
\name{ODRF.error}
\alias{ODRF.error}
\title{carPPRF Prediction}
\usage{
ODRF.error(forest, y, Xnew = NULL, ynew = NULL, ...)
}
\arguments{
\item{Xnew}{an n by d numeric matrix (preferable) or Xnew frame. The rows correspond to observations and columns correspond to features.}
}
\value{
carPPtree Prediction v_subs PPtree
}
\description{
Prediction a decision tree based on an input matrix and class vector.  This is the main function in the ppRF1 package.
}
\examples{
### Train RerF on numeric Xnew ###
library(rerf)
forest <- RerF1(as.matrix(iris[, 1:4]), iris[[5L]], num.cores = 1L)

### Train RerF on one-of-K encoded categorical Xnew ###
df1 <- as.Xnew.frame(Titanic)
nc <- ncol(df1)
df2 <- df1[NULL, -nc]
for (i in which(df1$Freq != 0L)) {
  df2 <- rbind(df2, df1[rep(i, df1$Freq[i]), -nc])
}
}
