% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODRF.R
\name{ODRF}
\alias{ODRF}
\title{ODRF forest Generator}
\usage{
ODRF(
  formula,
  data,
  subset = NULL,
  weights = NULL,
  na.action = na.fail,
  method = "i-classification",
  NodeRotateFun = "RotMatPPO",
  FunDir = getwd(),
  paramList = NULL,
  catLabel = NULL,
  Xcat = 0L,
  MinLeaf = ifelse(method == "regression", 5, 1),
  storeOOB = TRUE,
  TreeRandRotate = FALSE,
  replacement = TRUE,
  stratify = TRUE,
  ntrees = 100,
  numOOB = c(0, ceiling(length(y)/3))[2],
  numCores = 0L,
  seed = 220924,
  MaxDepth = Inf,
  numNode = Inf,
  Xscale = c("Min-max", "Quantile", "No")[1]
)
}
\arguments{
\item{method}{: the criterion used for splitting the nodes
'g-classification' : gini impurity index (classification)
'i-classification' : information gain (classification)
'regression' : squared error (regression)}

\item{ntrees}{: number of trees in the ensemble (default 50).}

\item{oobe}{: out-of-bag error calculation, values ('y'/'n' -> yes/no) (default 'n').}

\item{nsamtosample}{: number of randomly selected (withreplacement) samples to use to grow each tree (default num_samples)


Furthermore the following parameters can be set regarding the trees themselves :}

\item{minparent}{: the minimum amount of samples in an impure node for it to be considered for splitting}

\item{minleaf}{: the minimum amount of samples in a leaf}

\item{Xweights}{: a vector of values which weigh the samples when considering a split}

\item{nvartosample}{: the number of (randomly selected) variables to consider at each node}
}
\value{
ppForest
}
\description{
Creates an ensemble of carPPtrees using X(samplesXfeatures).
}
\details{
The following parameters can be set :
}
\examples{
### Train RerF on numeric data ###
library(rerf)
forest <- RerF1(as.matrix(iris[, 1:4]), iris[[5L]], num.cores = 1L)

}
