% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ODT.R
\name{plot.ODT}
\alias{plot.ODT}
\title{PPtree plot}
\usage{
\method{plot}{ODT}(
  tree,
  font.size = 17,
  width.size = 1,
  xadj = 0,
  main = paste0("Oblique ", ifelse(tree$method == "regression", "Regression",
    "Classification"), " Tree"),
  sub = NULL,
  ...
)
}
\arguments{
\item{font.size}{font size of plot}

\item{width.size}{size of eclipse in each node.}

\item{main}{main title}

\item{sub}{sub title}

\item{...}{arguments to be passed to methods}

\item{x}{PPtreeclass object}
}
\description{
projection pursuit classification tree plot
}
\details{
Draw projection pursuit classification tree with tree structure. It is 
modified from a function in party library.
}
\examples{
data(iris)
Tree.result <- PPTreeclass(Species~., data = iris,"LDA")
Tree.result
plot(Tree.result,xjust=3)
}
\references{
Lee, EK(2017) 
PPtreeViz: An R Package for Visualizing Projection Pursuit Classification 
Trees, Journal of Statistical Software <doi:10.18637/jss.v083.i08>
}
\keyword{tree}
