% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prune_ODT.R
\name{plot.prune.ODT}
\alias{plot.prune.ODT}
\title{prune oblique decision tree plot}
\usage{
\method{plot}{prune.ODT}(
  ppTree,
  position = "topleft",
  main = paste0("Oblique ", ifelse(ppTree$type == "regression", "Regression",
    "Classification"), " Tree"),
  ...
)
}
\arguments{
\item{ppTree}{an object of class \code{\link{prune.ODT}}.}

\item{position}{Position of the curve label.}

\item{main}{main title}

\item{...}{arguments to be passed to methods.}
}
\value{
Error of test data after each pruning, classification error rate for classification or RPE(MSE/mean((ytest-mean(y))^2)) for regression.
}
\description{
Draw the error graph of class \code{ODT} at different number of split nodes.
}
\examples{
library(ODRF)

data(seeds)
set.seed(221212)
train = sample(1:209,100)
train_data = data.frame(seeds[train,])
test_data = data.frame(seeds[-train,])

tree = ODT(varieties_of_wheat~.,train_data,type='i-classification')
tree = prune(tree,train_data)
#oblique decision tree plot (default)
plot(tree)
#prune oblique decision tree plot
class(tree)="prune.ODT"
plot(tree)

}
\seealso{
\code{ODT} \code{prune.ODRF}
}
\keyword{tree}
