% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_VarImp.R
\name{plot.VarImp}
\alias{plot.VarImp}
\title{Variable Importance Plot}
\usage{
\method{plot}{VarImp}(
  varImp,
  nvar = min(30, nrow(varImp$varImp)),
  main = paste0("Oblique ", ifelse(varImp$type == "regression", "Regression",
    "Classification"), " Forest"),
  ...
)
}
\arguments{
\item{varImp}{an object of class \code{\link{VarImp}}.}

\item{nvar}{How many variables to show?}

\item{main}{plot title.}

\item{...}{arguments to be passed to methods.}
}
\value{
A matrix of importance measure, first column for each predictor variable and second column is Increased error. 
classification error rate for classification or RPE(MSE/mean((ytest-mean(y))^2)) for regression.
}
\description{
Dotchart of variable importance as measured by a Oblique Decision Random Forest.
}
\examples{
data(breast_cancer)
forest = ODRF(diagnosis~.,seeds,type='i-classification')
varimp=VarImp(forest,seeds)
plot(varimp)

}
\seealso{
\code{\link{ODRF}} \code{\link{VarImp}}
}
