% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPO.R
\name{PPO}
\alias{PPO}
\title{Projection pursuit optimization}
\usage{
PPO(
  x,
  y,
  q = 1L,
  ppMethod = "LDA",
  weight = TRUE,
  r = 1L,
  lambda = 0.1,
  energy = 0,
  cooling = 0.9,
  TOL = 1e-04,
  maxiter = 1000L,
  ...
)
}
\arguments{
\item{y}{class information vector}

\item{q}{dimension of projection matrix}

\item{ppMethod}{method for projection pursuit; "LDA", "PDA", "Lr", "GINI", 
and "ENTROPY"}

\item{weight}{weight flag in LDA, PDA and Lr index}

\item{r}{r in Lr index}

\item{lambda}{lambda in PDA index}

\item{energy}{energy parameter}

\item{cooling}{cooling parameter}

\item{TOL}{tolerance}

\item{maxiter}{number of maximum iteration}

\item{X}{data matrix without class information}
}
\value{
bestIndex maximum LDA index value

projMat optimal q-dim projection matrix
}
\description{
PP optimization using various projection pursuit indices
}
\details{
Find the q-dim optimal projection using various projectin pursuit indices 
with class information
}
\examples{
data(iris)
PP.proj.result <- ppOpt(as.matrix(iris[,1:4]),iris[,5])
PP.proj.result
}
\references{
Lee, EK., Cook, D., Klinke, S., and Lumley, T.(2005) 
Projection Pursuit for exploratory supervised classification, 
Journal of Computational and Graphical statistics, 14(4):831-846.

Lee, YD, Cook, D., Park JW, and Lee, EK(2013) 
PPtree: Projection Pursuit Classification Tree, 
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{projection}
\keyword{pursuit}
