% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ODRF.R
\name{print.ODRF}
\alias{print.ODRF}
\title{Print PP.Tree.class result}
\usage{
\method{print}{ODRF}(forest, ...)
}
\arguments{
\item{x}{PPtreeclass object}
}
\description{
Print the projection pursuit classification tree result
}
\keyword{tree}
