% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ODRF.R
\name{print.ODRF}
\alias{print.ODRF}
\title{print ODRF}
\usage{
\method{print}{ODRF}(ppForest, ...)
}
\arguments{
\item{ppForest}{An object of class \code{\link{ODRF}}.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Print contents of ODRF object.
}
\examples{
data(iris)
forest <- ODRF(Species ~ ., data = iris, parallel = FALSE)
forest

}
\seealso{
\code{\link{ODRF}}
}
\keyword{forest}
