% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ODT.R
\name{print.ODT}
\alias{print.ODT}
\title{Print PP.Tree.class result}
\usage{
\method{print}{ODT}(tree, projection = FALSE, cutvalue = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{verbose}{print if TRUE, no output if FALSE}

\item{...}{arguments to be passed to methods}

\item{x}{PPtreeclass object}

\item{coef.print}{print projection coefficients in each node ifTRUE}

\item{cutoff.print}{print cutoff values in each node if TRUE}
}
\description{
Print PP.Tree.class result
}
\details{
Print the projection pursuit classification tree result
}
\examples{
data(iris)
Tree.result <- PPTreeclass(Species~.,data = iris,"LDA")
Tree.result
print(Tree.result,coef.print=TRUE,cutoff.print=TRUE)tree
}
\references{
Lee, EK(2017) 
PPtreeViz: An R Package for Visualizing Projection Pursuit Classification 
Trees, Journal of Statistical Software <doi:10.18637/jss.v083.i08>
}
\keyword{tree}
