% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ODT.R
\name{print.ODT}
\alias{print.ODT}
\title{print ODT result}
\usage{
\method{print}{ODT}(x, projection = FALSE, cutvalue = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{\link{ODT}}.}

\item{projection}{Print projection coefficients in each node if TRUE.}

\item{cutvalue}{Print cutoff values in each node if TRUE.}

\item{verbose}{Print if TRUE, no output if FALSE.}

\item{...}{Arguments to be passed to methods.}
}
\value{
The oblique decision tree structure.
}
\description{
Print the oblique decision tree structure.
}
\examples{
data(iris)
tree <- ODT(Species ~ ., data = iris)
tree
print(tree, projection = TRUE, cutvalue = TRUE)

}
\references{
Lee, EK(2017)
PPtreeViz: An R Package for Visualizing Projection Pursuit Classification
Trees, Journal of Statistical Software.
}
\seealso{
\code{\link{ODT}}
}
\keyword{print}
\keyword{tree}
