% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_ODT.R
\name{prune.ODT}
\alias{prune.ODT}
\title{RerF Tree Generator}
\usage{
\method{prune}{ODT}(ppTree, data, weights = NULL, MaxDepth = NULL)
}
\arguments{
\item{'g-classification'}{: gini impurity index (classification)}

\item{'i-classification'}{: information gain (classification, default)}

\item{'regression'}{: squared error (regression)

Other parameters that can be set are:}

\item{paramList}{parameters in a named list to be used by FUN. If left unchanged,}

\item{catMap}{a list specifying which columns in X correspond to the same one-of-K encoded feature. Each element of catMap is a numeric vector specifying the K column indices of X corresponding to the same categorical feature after one-of-K encoding. All one-of-K encoded features in X must come after the numeric features. The K encoded columns corresponding to the same categorical feature must be placed contiguously within X. The reason for specifying catMap is to adjust for the fact that one-of-K encoding cateogorical features results in a dilution of numeric features, since a single categorical feature is expanded to K binary features. If catMap = NULL, then RerF assumes all features are numeric (i.e. none of the features have been one-of-K encoded).}

\item{FUN}{:FUN=RandMatPPR.}

\item{minparent}{: the minimum amount of samples in an impure node for it to be considered for splitting (default 2)}

\item{MinLeaf}{: the minimum amount of samples in a leaf (default 1)}

\item{Xweights}{: a vector of values which weigh the samples when considering a split (default [])}

\item{nvartosample}{: the number of (randomly selected) variables to consider at each node (default all)}
}
\value{
ppTree
}
\description{
Grows a CARTree using X(samplesXfeatures) and one of the following criteria which can be set via the parameter 'method'
}
\examples{
### Train RerF on numeric data ###
library(rerf)
forest <- RerF1(as.matrix(iris[, 1:4]), iris[[5L]], num.cores = 1L)

}
