% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_ODT.R
\name{prune.ODT}
\alias{prune.ODT}
\title{Pruning of class \code{ODT}.}
\usage{
\method{prune}{ODT}(ppTree, X, y, MaxDepth = 1, ...)
}
\arguments{
\item{ppTree}{an object of class \code{ODT}.}

\item{X}{An n by d numeric matrix (preferable) or data frame is used to prune the object of class \code{ODT}.}

\item{y}{A response vector of length n.}

\item{MaxDepth}{The maximum depth of the tree after pruning. (Default 1)}

\item{...}{optional parameters to be passed to the low level function.}
}
\value{
\itemize{an object of class \code{ODT} and \code{prune.ODT}. 
\item{\code{ppTree} The same result as \code{ODT}.}
\item{\code{pruneError} Error of validation data after each pruning, misclassification rate (MR) for classification or mean square error (MSE) for regression.}
}
}
\description{
Prune \code{ODT} from bottom to top with validation data based on prediction error.
}
\examples{
#Classification with Oblique Decision Tree
data(seeds)
set.seed(221212)
train = sample(1:209,100)
train_data = data.frame(seeds[train,])
test_data = data.frame(seeds[-train,])
index=seq(floor(nrow(train_data)/2))

tree = ODT(varieties_of_wheat~.,train_data[index,],type='i-classification')
tree = prune(tree,train_data[-index,-8],train_data[-index,8])
pred <- predict(tree,test_data[,-8])
#classification error
(mean(pred!=test_data[,8]))


#Regression with Oblique Decision Tree
data(body_fat)
set.seed(221212)
train = sample(1:252,100)
train_data = data.frame(body_fat[train,])
test_data = data.frame(body_fat[-train,])
index=seq(floor(nrow(train_data)/2))

tree = ODT(Density~.,train_data[index,],type='regression')
tree = prune(tree,train_data[-index,-1],train_data[-index,1])
pred <- predict(tree,test_data[,-1])
#estimation error
mean((pred-test_data[,1])^2)

}
\seealso{
\code{\link{ODT}} \code{\link{plot.prune.ODT}} \code{\link{prune.ODRF}} \code{\link{online.ODT}}
}
