% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotMat.R
\name{RotMatPPO}
\alias{RotMatPPO}
\title{Create a Random Matrix: RandMatPPR}
\usage{
RotMatPPO(
  x,
  y,
  numProj,
  dimProj,
  catLabel = NULL,
  ppMethod = "PPR",
  method = "i-classification",
  weights = NULL
)
}
\arguments{
\item{x}{an n by d numeric matrix (preferable) or data frame. The rows correspond to observations and columns correspond to features.}

\item{y}{an n length vector of class labels.  Class labels must be integer or numeric and be within the range 1 to the number of classes.}

\item{catLabel}{a list specifying specifies which one-of-K encoded columns in X correspond to the same categorical feature.}

\item{d}{the number of desired columns in the projection matrix.}

\item{...}{used to handle superfluous arguments passed in using paramList.}
}
\value{
A random matrix to use in running \code{\link{RerF}}.
}
\description{
Create a Random Matrix: RandMatPPR
}
\examples{

x <- matrix(rnorm(200),20,10)
y <- (rnorm(20)>0)+0
d <- 3
sparsity <- 0.25
set.seed(220828)
(a <- RandMatPPR(x,y,d,sparsity,catMap=NULL))
}
