% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotMat.R
\name{RotMatRand}
\alias{RotMatRand}
\title{Random Rotation Matrix}
\usage{
RotMatRand(
  dimX,
  randDist = "Binary",
  numProj = ceiling(sqrt(dimX)),
  dimProj = "Rand",
  sparsity = ifelse(dimX >= 10, 3/dimX, 1/dimX),
  prob = 0.5,
  lambda = 1,
  catLabel = NULL,
  ...
)
}
\arguments{
\item{dimX}{the number of dimensions.}

\item{randDist}{The probability distribution of the random projection direction from. "Binary", "Norm", "Uniform" denote the B{-1,1} 
binomial distribution (default), N(0,1) normal distribution and U(-1,1) uniform distribution, respectively.}

\item{numProj}{the number of desired columns in the projection matrix, is also the number of projection directions.(default ceiling(sqrt(dimX)))}

\item{dimProj}{Number of variables to be projected, default dimProj="Rand": random from 1 to ncol(X).}

\item{sparsity}{a real number in \eqn{(0,1)} that specifies the distribution of non-zero elements in the random matrix. 
When sparsity="pois" means that non-zero elements are generated by the p(\code{lambda}) Poisson distribution.}

\item{prob}{a probability \eqn{\in (0,1)} used for sampling from.
\eqn{{-1,1}} where \code{prob = 0} will only sample -1 and \code{prob = 1} will only sample 1.}

\item{lambda}{Parameter of the Poisson distribution (default 1).}

\item{catLabel}{A category labels of class \code{list} in prediction variables，for details see Examples of \code{\link{ODRF}}.}

\item{...}{used to handle superfluous arguments passed in using paramList.}
}
\value{
\itemize{A random matrix to use in running \code{\link{ODT}}.
\item{Variable: variables to be projected.} 
\item{Number: Number of variables to be projected.}
\item{Coefficient: Coefficients of the projection matrix.}
}
}
\description{
Generate various rotation matrices by different distributions.
}
\examples{
set.seed(1)
paramList <- list(dimX = 8, numProj= 3, sparsity=0.25, prob=0.5)
(RotMat <- do.call(RotMatRand, paramList))
paramList <- list(dimX = 8,numProj= 3,sparsity="pois")
(RotMat <- do.call(RotMatRand, paramList))
paramList <- list(dimX = 8, randDist="Norm", dimProj=5)
(RotMat <- do.call(RotMatRand, paramList))
}
\seealso{
\code{\link{RotMatPPO}} \code{\link{RotMatRF}} \code{\link{RotMatMake}}
}
